src/api/pyproject.toml
[project]
name = "kartograph-api"
version = "3.26.0"
description = "Enterprise-Ready Bi-Temporal Knowledge Graphs as a Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "apache-age-python>=0.0.7",
    "asyncpg>=0.31.0",
    "asyncpg-listen>=0.0.9",
    "authzed>=1.24.0",
    "bcrypt>=5.0.0",
    "cyclopts==5.0.0a1",
    "dlt>=1.23.0",
    "fastapi[standard]>=0.123.9",
    "fastmcp==2.14.3",
    "httpx>=0.28.1",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-ulid>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-archon>=0.0.7",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.2.0",
    "pytest-mock>=3.15.1",
    "rich>=14.2.0",
    "ruff>=0.14.8",
    "types-psycopg2>=2.9.21.20251012",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
pythonpath = ["."]

[tool.pytest_env]
# Database settings for integration tests
# Use skip_if_set = true so worktree-dev.sh port overrides are respected
KARTOGRAPH_DB_HOST = { value = "localhost", skip_if_set = true }
KARTOGRAPH_DB_PORT = { value = "5432", skip_if_set = true }
KARTOGRAPH_DB_DATABASE = { value = "kartograph", skip_if_set = true }
KARTOGRAPH_DB_USERNAME = { value = "kartograph", skip_if_set = true }
KARTOGRAPH_DB_PASSWORD = { value = "kartograph_dev_password", skip_if_set = true }

# SpiceDB settings for integration tests
SPICEDB_ENDPOINT = { value = "localhost:50051", skip_if_set = true }
SPICEDB_PRESHARED_KEY = { value = "changeme", skip_if_set = true }
SPICEDB_USE_TLS = { value = "true", skip_if_set = true }
# Path to self-signed certificate (relative to src/api)
SPICEDB_CERT_PATH = { value = "../../certs/spicedb-cert.pem", transform = true, skip_if_set = true }
